requirements.txt
flask
pymongo
spark
pandas
json
plotly